requirements.txt
# Only need dateutil for date parsing
python-dateutil==2.8.2
Flask==3.0.2 